notice 타입 정리
